refactor(grpc): move Qwen3-ASR family knowledge out of the router
The gRPC router's transcription endpoint had ~250 lines of Qwen3-ASR
specifics inlined: a 30-language table, model-id/label substring
detection, prompt sanitization, the family chat-request shape, and
output parsing -- a whole model family owned by the router layer, and a
second source of family truth that could drift from the multimodal
crate's qwen3_asr spec (the two detected the family by copy-pasted
substring checks).

Put the family knowledge where the family already lives. crates/multimodal
qwen3_asr grows a protocol-free `transcription` module -- identifiers,
the language set, prompt sanitation, the asr_text framing, output
parsing -- with plain-data errors (PromptTooLong, UnsupportedLanguage)
for a caller to render, and the audio-processing spec's own matches()
now calls the shared identifier check instead of duplicating it.

The gateway keeps only what is genuinely its layer: a
TranscriptionModelSpec trait + a first-match registry, and a Qwen3-ASR
adapter that maps the shared family knowledge onto the HTTP endpoint
(worker-label detection, the chat request, capability gating, error
codes). route_audio_transcriptions is now family-blind -- resolve a
spec or 400, then validate/build/execute/render through the trait.
Adding another transcription family is a new adapter module and one
registry row, not another conditional in router.rs.

No behavior change: same detection, same request shape, same error
codes and bodies, same output parsing. router.rs drops ~440 lines; the
moved tests split to sit beside the code they cover, plus a new adapter
test that the family error variants map to the right HTTP codes.
