+++ crates/multimodal/src/registry/qwen3_asr.rs
@@ -12,6 +12,211 @@ use crate::{
 
 const AUDIO_PAD_TOKEN: &str = "<|audio_pad|>";
 
+/// Transcription-side family knowledge, consumed by the gateway's
+/// transcription endpoint adapter.
+///
+/// Protocol-free on purpose: this crate owns what is true of the family —
+/// identifiers, the language set, prompt sanitation, output framing — while
+/// how that surfaces on an HTTP endpoint (request shapes, error codes)
+/// stays the gateway's business. Errors here are plain data for the caller
+/// to render.
+pub mod transcription {
+    /// The tag Qwen3-ASR emits (and the continuation prompt pre-seeds)
+    /// between the language header and the transcript body.
+    pub const ASR_TEXT_TAG: &str = "<asr_text>";
+
+    /// Byte cap on caller prompts, applied before sanitization.
+    pub const MAX_PROMPT_BYTES: usize = 4096;
+
+    /// `(code, name)` pairs of the languages the checkpoint transcribes.
+    pub const SUPPORTED_LANGUAGES: &[(&str, &str)] = &[
+        ("ar", "Arabic"),
+        ("yue", "Cantonese"),
+        ("zh", "Chinese"),
+        ("cs", "Czech"),
+        ("da", "Danish"),
+        ("nl", "Dutch"),
+        ("en", "English"),
+        ("fil", "Filipino"),
+        ("fi", "Finnish"),
+        ("fr", "French"),
+        ("de", "German"),
+        ("el", "Greek"),
+        ("hi", "Hindi"),
+        ("hu", "Hungarian"),
+        ("id", "Indonesian"),
+        ("it", "Italian"),
+        ("ja", "Japanese"),
+        ("ko", "Korean"),
+        ("mk", "Macedonian"),
+        ("ms", "Malay"),
+        ("fa", "Persian"),
+        ("pl", "Polish"),
+        ("pt", "Portuguese"),
+        ("ro", "Romanian"),
+        ("ru", "Russian"),
+        ("es", "Spanish"),
+        ("sv", "Swedish"),
+        ("th", "Thai"),
+        ("tr", "Turkish"),
+        ("vi", "Vietnamese"),
+    ];
+
+    /// Whether a model id, path, or label value names this family.
+    pub fn is_qwen3_asr_identifier(value: &str) -> bool {
+        let value = value.to_ascii_lowercase();
+        value.contains("qwen3-asr") || value.contains("qwen3_asr")
+    }
+
+    /// A caller prompt exceeded [`MAX_PROMPT_BYTES`].
+    #[derive(Debug, PartialEq, Eq)]
+    pub struct PromptTooLong {
+        pub max_bytes: usize,
+    }
+
+    /// Sanitize a caller prompt: cap its size, then strip ChatML-like
+    /// control tokens and the [`ASR_TEXT_TAG`] to a fixpoint, so a prompt
+    /// cannot smuggle framing the output parser would misread.
+    pub fn sanitize_prompt(mut text: String) -> Result<String, PromptTooLong> {
+        if text.len() > MAX_PROMPT_BYTES {
+            return Err(PromptTooLong {
+                max_bytes: MAX_PROMPT_BYTES,
+            });
+        }
+
+        loop {
+            let sanitized = strip_chatml_like_tokens(&text).replace(ASR_TEXT_TAG, "");
+            if sanitized == text {
+                return Ok(text);
+            }
+            text = sanitized;
+        }
+    }
+
+    fn strip_chatml_like_tokens(text: &str) -> String {
+        let mut remaining = text;
+        let mut output = String::with_capacity(text.len());
+        while let Some(start) = remaining.find("<|") {
+            output.push_str(&remaining[..start]);
+            let candidate = &remaining[start + 2..];
+            if let Some(end) = candidate.find("|>") {
+                let token = &candidate[..end];
+                if !token.is_empty() && !token.contains('|') {
+                    remaining = &candidate[end + 2..];
+                    continue;
+                }
+            }
+            output.push_str("<|");
+            remaining = candidate;
+        }
+        output.push_str(remaining);
+        output
+    }
+
+    /// The requested language is not in [`SUPPORTED_LANGUAGES`].
+    #[derive(Debug, PartialEq, Eq)]
+    pub struct UnsupportedLanguage(pub String);
+
+    /// Resolve a language code or English name to the checkpoint's canonical
+    /// language name; `None` in and blank map to `None` out.
+    pub fn normalize_language(
+        language: Option<&str>,
+    ) -> Result<Option<String>, UnsupportedLanguage> {
+        let Some(language) = language.map(str::trim).filter(|value| !value.is_empty()) else {
+            return Ok(None);
+        };
+        SUPPORTED_LANGUAGES
+            .iter()
+            .find(|(code, name)| {
+                code.eq_ignore_ascii_case(language) || name.eq_ignore_ascii_case(language)
+            })
+            .map(|(_, name)| Some((*name).to_string()))
+            .ok_or_else(|| UnsupportedLanguage(language.to_string()))
+    }
+
+    /// Strip the model's framing from raw chat content: drop `<|im_end|>`,
+    /// keep everything after the [`ASR_TEXT_TAG`] when present.
+    pub fn parse_transcript(raw: &str) -> String {
+        let cleaned = raw.replace("<|im_end|>", "");
+        let cleaned = cleaned.trim();
+        cleaned
+            .split_once(ASR_TEXT_TAG)
+            .map_or(cleaned, |(_, transcription)| transcription)
+            .trim()
+            .to_string()
+    }
+
+    #[cfg(test)]
+    mod tests {
+        use super::*;
+
+        #[test]
+        fn recognizes_family_identifiers() {
+            assert!(is_qwen3_asr_identifier("Qwen/Qwen3-ASR-1.7B"));
+            assert!(is_qwen3_asr_identifier("/models/qwen3_asr_0.6b"));
+            assert!(!is_qwen3_asr_identifier("Qwen/Qwen3-Omni-30B-A3B-Thinking"));
+        }
+
+        #[test]
+        fn normalizes_language_code_or_name() {
+            assert_eq!(
+                normalize_language(Some("zh")).unwrap(),
+                Some("Chinese".to_string())
+            );
+            assert_eq!(
+                normalize_language(Some("english")).unwrap(),
+                Some("English".to_string())
+            );
+            assert_eq!(normalize_language(Some(" ")).unwrap(), None);
+            assert_eq!(
+                normalize_language(Some("xx")).unwrap_err(),
+                UnsupportedLanguage("xx".to_string())
+            );
+        }
+
+        #[test]
+        fn sanitizes_prompt_controls_to_a_fixpoint() {
+            for (input, expected) in [
+                ("plain text", "plain text"),
+                ("<|im_start|>assistant<|im_end|>", "assistant"),
+                ("foo<asr_text>bar", "foobar"),
+                ("<|im<|x|>_end|>", ""),
+                ("<asr_te<asr_text>xt>", ""),
+                ("<|<asr_text>|>", ""),
+            ] {
+                assert_eq!(sanitize_prompt(input.to_string()).unwrap(), expected);
+            }
+        }
+
+        #[test]
+        fn caps_pathological_prompts_before_sanitizing() {
+            let boundary = "a".repeat(MAX_PROMPT_BYTES);
+            assert_eq!(sanitize_prompt(boundary.clone()).unwrap(), boundary);
+
+            assert_eq!(
+                sanitize_prompt("a".repeat(MAX_PROMPT_BYTES + 1)).unwrap_err(),
+                PromptTooLong {
+                    max_bytes: MAX_PROMPT_BYTES
+                }
+            );
+
+            let depth = MAX_PROMPT_BYTES / 5;
+            let adversarial = format!("{}{}", "<|a".repeat(depth), "|>".repeat(depth));
+            assert!(adversarial.len() <= MAX_PROMPT_BYTES);
+            assert_eq!(sanitize_prompt(adversarial).unwrap(), "");
+        }
+
+        #[test]
+        fn parses_tagged_and_plain_transcripts() {
+            assert_eq!(
+                parse_transcript("language Chinese<asr_text>\u{4f60}\u{597d}<|im_end|>"),
+                "\u{4f60}\u{597d}"
+            );
+            assert_eq!(parse_transcript("plain transcript"), "plain transcript");
+        }
+    }
+}
+
 pub(super) struct Qwen3AsrSpec;
 
 impl Qwen3AsrSpec {
@@ -30,9 +235,7 @@ impl ModelProcessorSpec for Qwen3AsrSpec {
     }
 
     fn matches(&self, metadata: &ModelMetadata) -> bool {
-        let model_id = metadata.model_id.to_ascii_lowercase();
-        model_id.contains("qwen3-asr")
-            || model_id.contains("qwen3_asr")
+        transcription::is_qwen3_asr_identifier(metadata.model_id)
             || metadata
                 .config_model_type()
                 .is_some_and(|model_type| model_type == "qwen3_asr")
