+++ model_gateway/src/routers/grpc/mod.rs
@@ -17,6 +17,7 @@ pub(crate) mod proto_wrapper;
 pub(crate) mod regular;
 pub(crate) mod router; // Used by routers/factory
 pub(crate) mod spec;
+pub(crate) mod transcription;
 pub mod utils; // Used by routers/http and bindings/golang
 pub mod zmq_client; // ZMQ backend adapter behind the vLLM client surface
 pub(crate) mod zmq_multimodal; // Proto mm inputs → EngineCore mm_features
